moved SearchField component to helper-components folder
