+++ client/src/components/add-entries/helper-components/SearchField.js
